Added math / display functions
calc.js was updated with 4 new mathematical funcitons(add/sub/mult/division) which take in two parameters and returns the corresponding expressions. A operator expression was added as well, which works the same as the last 4 but has a third parameter which takes in an operator to perform on the num inputs. Lastly, a nodelist was created for the button on the face of the calculator to display whichever button was clicked on.