Operation implementation
At this point in the program life I placed the numbers/operation buttons clicked by the user in global variables so I can use the operator function to run and return their result. However, I am running into an issue where the result is getting returned before the second number of the expression  is clicked.