Adjusted logic for operation return
In an attempt to fix the previous bug described, I moved the declaration of num2 to only happen if num1 is already assigned a value. I also create a bool variable to change from false to true when the equal button is clicked. This is so the operator function wont run until the equal button is clicked. However, I ran into another problem since 'text' was declared inside the else statement I am getting a reference error. So I am going to change 'text' and number variables to global, so when a button is clicked they are just updated.