Fixed click bug
Before patch, the user would have to click the around the operator symbol for the event listener to fire. This was because the image is a child to the parent element who holds the class that the event listener listens for. So the image itself was skipped over because it does not have the button class. So to fix it, I completly removed the parent element and just added the button class to the img element.